flask-compare-service/requirements.txt
Flask==2.2.5
face_recognition==1.3.0
numpy==1.25.0
opencv-python-headless==4.8.0.74
Pillow==10.0.0
gunicorn==20.1.0
